requirements.txt
certifi==2020.6.20
cycler==0.10.0
kiwisolver==1.3.1
matplotlib==3.3.2
numpy==1.19.4
opencv-contrib-python==4.4.0.46
opencv-python==4.4.0.46
opencv-python3==1.0
Pillow==8.0.1
pkg-resources==0.0.0
pyparsing==2.4.7
python-dateutil==2.8.1
six==1.15.0
